Add fetcherUser for MySQL, refactoring
